boot_control: Do not swap guid's for the xbl_config partition
Selection of the appropriate a/b version of xbl_config is handled
by the bootloader based on the active LUN being booted from. We do
not need to swap GUID's for this partition.

We also now correctly return a error when setting the active slot on
one of the LUN's fails.

Change-Id: Idad9d0aa3b3e2e1bd29c90414b06898e967abb8a
Signed-off-by: Joshua Blanchard <joshua.lee.bbg@gmail.com>